+++ app/Models/Student.php
@@ -23,4 +23,9 @@ class Student extends Model
     {
         return $this->belongsTo(School::class, 'school_id');
     }
+
+    public function attendances()
+    {
+        return $this->hasMany(Attendance::class, 'student_id');
+    }
 }